payment/requirements.txt
fastapi[all]
stripe
python-dotenv